requirements.txt
%%writefile hf_space_files/requirements.txt
torch
torchvision
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
faiss-cpu                  # use faiss-gpu-cu12 if you enable GPU in Space
pillow
numpy
gradio>=4.0.0